Move units and plot preferences to top level map.
