+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/preferences/CoPlotManagementWindow.java
@@ -77,7 +77,7 @@ public class CoPlotManagementWindow extends javax.swing.JInternalFrame {
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({ "unchecked", "rawtypes" })
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
         java.awt.GridBagConstraints gridBagConstraints;
