+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/preferences/SedModel.java
@@ -25,13 +25,9 @@ import java.util.Map.Entry;
 import org.apache.commons.lang.StringUtils;
 
 import cfa.vo.iris.sed.ExtSed;
-import cfa.vo.iris.sed.SedException;
-import cfa.vo.iris.sed.quantities.SPVYQuantity;
-import cfa.vo.iris.sed.quantities.SPVYUnit;
 import cfa.vo.iris.visualizer.plotter.ColorPalette;
 import cfa.vo.iris.visualizer.plotter.HSVColorPalette;
 import cfa.vo.iris.units.UnitsException;
-import cfa.vo.iris.visualizer.plotter.PlotPreferences;
 import cfa.vo.iris.visualizer.stil.tables.IrisStarTable;
 import cfa.vo.iris.visualizer.stil.tables.IrisStarTableAdapter;
 import cfa.vo.sedlib.Segment;
@@ -50,7 +46,6 @@ public class SedModel {
     final Map<Segment, LayerModel> segmentModels;
     final ExtSed sed;
     final ColorPalette colors;
-    final PlotPreferences plotPreferences;
     
     private String xunits;
     private String yunits;
@@ -59,7 +54,6 @@ public class SedModel {
         this.sed = sed;
         this.adapter = adapter;
         this.colors = new HSVColorPalette();
-        this.plotPreferences = PlotPreferences.getDefaultPlotPreferences();
 
         this.segmentModels = Collections.synchronizedMap(new IdentityHashMap<Segment, LayerModel>());
         
@@ -173,8 +167,9 @@ public class SedModel {
     private void setUnits(Segment seg, LayerModel layer) {
         // if this is the first segment to be added to the SED preferences,
         // set the preferred X and Y units to the segment's
-        if (StringUtils.isEmpty(xunits) || StringUtils.isEmpty(xunits))
+        if (StringUtils.isEmpty(xunits) || StringUtils.isEmpty(xunits)) {
             setInitialUnits(seg);
+        }
         
         // set the layer units for this segment
         try {
@@ -199,13 +194,6 @@ public class SedModel {
         } catch (SedNoDataException ex) {
             throw new RuntimeException(ex);
         }
-        
-        // set the plot axes labels
-        plotPreferences.setXlabel(xunits);
-        plotPreferences.setYlabel(yunits);
-        
-        // if in magnitudes, flip the direction of the Y-axis
-        flipYifMag(yunits);
     }
     
     /**
@@ -217,8 +205,6 @@ public class SedModel {
     public void setUnits(String xunit, String yunit) {
         xunits = xunit;
         yunits = yunit;
-        plotPreferences.setXlabel(xunit);
-        plotPreferences.setYlabel(yunit);
         
         // update the segment layers with the new units
         for (LayerModel seg : segmentModels.values()) {
@@ -229,27 +215,6 @@ public class SedModel {
                 Logger.getLogger(SedModel.class.getName())
                         .log(Level.SEVERE, null, ex);
             }
-        }
-        // if in magnitudes, flip the direction of the Y-axis
-        flipYifMag(yunit);
-    }
-    
-    /**
-     * Flips the Y-axis if yunit is a magnitude. A lower magnitude = brighter 
-     * source  higher on Y-axis)
-     * @param yunit 
-     */
-    private void flipYifMag(String yunit) {
-        try {
-            if (SPVYQuantity.MAGNITUDE.getPossibleUnits()
-                    .contains(SPVYUnit.getFromUnitString(yunit))) {
-                plotPreferences.setYflip(true);
-            } else {
-                plotPreferences.setYflip(false);
-            }
-        } catch (SedException ex) {
-            Logger.getLogger(SedModel.class.getName())
-                    .log(Level.SEVERE, null, ex);
         }
     }
     
@@ -305,13 +270,5 @@ public class SedModel {
         for (LayerModel layer : segmentModels.values()) {
             layer.setYUnits(yunits);
         }
-    }
-    
-    /**
-     * @return
-     *  Top level plot preferences for the stil plotter.
-     */
-    public PlotPreferences getPlotPreferences() {
-        return plotPreferences;
     }
 }
